Create awards.test.js (closes #123)
